Create a review item component to isolate logic from the list
