+++ client/my-sites/marketplace/components/reviews-list/style.scss
@@ -1,4 +1,3 @@
-$profile-picture-size: 36px;
 $border-color: #eee;
 
 .marketplace-reviews-list__container {
@@ -9,109 +8,6 @@ $border-color: #eee;
 		color: var(--studio-gray-100);
 	}
 
-	.marketplace-reviews-list__customer-reviews {
-		.marketplace-reviews-list__review-container {
-			border-top: 1px solid $border-color;
-			padding: 32px 0;
-
-			&:first-of-type {
-				border-top: none;
-				padding-top: 0;
-			}
-
-			&.last {
-				padding-bottom: 0;
-			}
-
-			.marketplace-reviews-list__review-container-header {
-				display: grid;
-				grid-template-columns: calc($profile-picture-size + 8px) 1fr 100px;
-
-				.marketplace-reviews-list__rating-data {
-					display: flex;
-					flex-direction: column;
-
-					.rating {
-						margin-left: -1px;
-					}
-				}
-
-				.marketplace-reviews-list__profile-picture {
-					display: flex;
-					flex-direction: column;
-					justify-content: center;
-
-					.marketplace-reviews-list__profile-picture-img,
-					.marketplace-reviews-list__profile-picture-placeholder {
-						height: $profile-picture-size;
-						width: $profile-picture-size;
-						border-radius: 50%;
-					}
-
-					.marketplace-reviews-list__profile-picture-placeholder {
-						background-color: var(--studio-gray-20);
-					}
-				}
-
-				.marketplace-reviews-list__date {
-					color: var(--studio-gray-60);
-					font-size: $font-body-small;
-					text-align: right;
-					align-self: center;
-				}
-
-				.marketplace-reviews-list__author {
-					color: var(--studio-gray-100);
-					font-size: $font-body-small;
-				}
-			}
-
-			.marketplace-reviews-list__review-rating {
-				margin-top: 16px;
-				margin-bottom: 16px;
-				display: flex;
-				align-items: center;
-				justify-content: space-between;
-
-				h2 {
-					font-family: "SF Pro Text", $sans;
-					font-size: $font-title-small;
-					font-weight: 500;
-					line-height: 26px;
-					overflow-wrap: anywhere;
-					color: var(--studio-gray-100);
-				}
-			}
-
-			.marketplace-reviews-list__content {
-				color: var(--studio-gray-100);
-				font-size: $font-body-small;
-				margin-top: 16px;
-
-				p {
-					margin-bottom: 0;
-				}
-
-				&.edting {
-					background-color: var(--studio-gray-0);
-					padding: 16px;
-				}
-			}
-
-			.marketplace-reviews-list__editor {
-				color: var(--studio-gray-100);
-				background-color: var(--studio-gray-0);
-				border-radius: 4px;
-				font-size: $font-body-small;
-				margin-top: 16px;
-				padding: 16px;
-
-				p {
-					margin-bottom: 0;
-				}
-			}
-		}
-	}
 	.rating .rating__overlay .is-empty,
 	.rating .rating__star-outline .is-empty {
 		fill: var(--studio-yellow-20);
@@ -123,36 +19,6 @@ $border-color: #eee;
 	}
 }
 
-.marketplace-reviews-list__review-actions {
-	display: flex;
-	align-items: baseline;
-	flex-wrap: wrap;
-	justify-content: space-between;
-	margin-top: 15px;
-	font-size: $font-body-small;
-
-	.marketplace-reviews-list__review-actions-editable {
-		display: flex;
-		align-items: center;
-
-		.marketplace-reviews-list__review-submit {
-			margin-left: 16px;
-		}
-
-		.marketplace-reviews-list__review-actions-editable-button {
-			color: var(--studio-blue-50);
-			margin-right: 15px;
-			cursor: pointer;
-			display: flex;
-			align-items: center;
-		}
-
-		.gridicon {
-			margin-right: 5px;
-			color: var(--studio-gray-10);
-		}
-	}
-}
 
 .marketplace-reviews-list__no-reviews,
 .marketplace-reviews-list__customer-reviews {
@@ -182,23 +48,6 @@ $border-color: #eee;
 	}
 }
 
-.marketplace-reviews-list__pending-review,
-.marketplace-reviews-list__error-message {
-	display: flex;
-	align-items: flex-start;
-	gap: 8px;
-	align-self: stretch;
-
-	.marketplace-reviews-list__card-icon {
-		color: var(--studio-gray-10);
-	}
-
-	.marketplace-reviews-list__card-text {
-		font-family: "SF Pro Text", $sans;
-		font-size: $font-body-small;
-		line-height: 20px;
-	}
-}
 
 .confirm-modal {
 	z-index: z-index(".dialog__backdrop", ".confirm-modal");
